workers/analyst/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
structlog>=24.0.0
python-dotenv>=1.0.0
prometheus_client>=0.20.0
